# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DIGITALYCloud API",
    "version": "1.0.0",
    "description": "Hosting for Discord bots, Node.js apps, APIs and workers. All paths are relative to `/v1`. Timestamps are epoch milliseconds, dates `YYYY-MM-DD`, money in euros (integer cents internally). Errors: `{ \"error\": { \"code\", \"message\", \"fields?\" } }`. Session principals pick the team with `X-Team-Id` (or `?team=` on GET/EventSource); API keys are bound to one team. `x-permission` lists the team roles and the minimum API-key scope for each operation (null: API keys not accepted)."
  },
  "servers": [
    {
      "url": "https://api.cloud.digitaly.fr/v1"
    },
    {
      "url": "http://localhost:4000/v1"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "dgc_session",
        "description": "HttpOnly session cookie set by signup/login. Unsafe requests also need `X-CSRF-Token` (see GET /auth/csrf) and an allowed Origin."
      },
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "`Authorization: Bearer dgc_live_…` — team-bound, scope `read` or `full`, 120 requests/min."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "examples": [
                  "VALIDATION_ERROR",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "CSRF_TOKEN_INVALID",
                  "BILLING_UNAVAILABLE",
                  "PAYMENT_METHOD_REQUIRED"
                ]
              },
              "message": {
                "type": "string",
                "description": "User-facing message, shown as-is by the dashboard."
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                },
                "description": "Per-field validation messages."
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": true,
            "description": "Extra details depend on the code, e.g. `checkoutUrl`/`operationId` (402), `retryAfter` (429)."
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "CreateServiceInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[A-Za-z][\\w-]{1,31}$"
          },
          "type": {
            "type": "string",
            "enum": [
              "discord",
              "node",
              "api",
              "worker"
            ]
          },
          "source": {
            "type": "string",
            "enum": [
              "github",
              "upload",
              "docker"
            ]
          },
          "repo": {
            "type": "string",
            "description": "`owner/repo` (github) or an image reference (docker); empty for uploads."
          },
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "uploadId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "nodeVersion": {
            "type": "string",
            "enum": [
              "24 LTS",
              "22 LTS"
            ]
          },
          "startCommand": {
            "type": "string",
            "maxLength": 1000
          },
          "port": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1,
                "maximum": 65535
              },
              {
                "type": "null"
              }
            ]
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "starter",
              "pro",
              "business"
            ]
          },
          "regionId": {
            "type": "string"
          },
          "env": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvVarInput"
            }
          },
          "autoDeploy": {
            "type": "boolean"
          },
          "autoRestart": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "type",
          "source",
          "startCommand",
          "port",
          "plan",
          "regionId"
        ],
        "additionalProperties": false
      },
      "UpdateServiceInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startCommand": {
            "type": "string"
          },
          "nodeVersion": {
            "type": "string"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ]
          },
          "branch": {
            "type": "string"
          },
          "autoDeploy": {
            "type": "boolean"
          },
          "autoRestart": {
            "type": "boolean"
          }
        },
        "required": [],
        "additionalProperties": false
      },
...
      "ServiceDeploymentPair": {
        "type": "object",
        "properties": {
          "service": {
            "$ref": "#/components/schemas/Service"
          },
          "deployment": {
            "$ref": "#/components/schemas/Deployment"
          }
        },
        "required": [
          "service",
          "deployment"
        ],
        "additionalProperties": false
      },
...
      "ServiceList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "data",
          "nextCursor"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/services": {
      "get": {
        "summary": "List services",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "session": []
          },
          {
            "apiKey": []
          }
        ],
        "x-permission": {
          "action": "services.read",
          "roles": [
            "owner",
            "admin",
            "developer",
            "viewer"
          ],
          "apiKeyScope": "read"
        },
        "parameters": [
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Team for this request (session principals). Defaults to the session’s active team."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceList"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in / invalid key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (role, scope, CSRF, team)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a service and its first deployment",
        "description": "Paid plans charge the prorated amount now; without a card the owner gets `402` with `checkoutUrl` and `operationId`, and the service is created after payment.",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "x-permission": {
          "action": "services.create",
          "roles": [
            "owner",
            "admin"
          ],
          "apiKeyScope": null
        },
        "x-csrf": "Cookie-authenticated calls need Origin + X-CSRF-Token; bearer keys are exempt.",
        "parameters": [
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Team for this request (session principals). Defaults to the session’s active team."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9._-]{8,100}$"
            },
            "description": "Replays the first response for 24 h; a different body with the same key is 422."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceDeploymentPair"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in / invalid key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "Payment required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (role, scope, CSRF, team)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Unavailable / not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}": {
      "get": {
        "summary": "Get a service",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "session": []
          },
          {
            "apiKey": []
          }
        ],
        "x-permission": {
          "action": "services.read",
          "roles": [
            "owner",
            "admin",
            "developer",
            "viewer"
          ],
          "apiKeyScope": "read"
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Team for this request (session principals). Defaults to the session’s active team."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in / invalid key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (role, scope, CSRF, team)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update settings (applied by the next deployment)",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "session": []
          },
          {
            "apiKey": []
          }
        ],
        "x-permission": {
          "action": "services.update",
          "roles": [
            "owner",
            "admin"
          ],
          "apiKeyScope": "full"
        },
        "x-csrf": "Cookie-authenticated calls need Origin + X-CSRF-Token; bearer keys are exempt.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Team for this request (session principals). Defaults to the session’s active team."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in / invalid key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (role, scope, CSRF, team)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete (asynchronous purge; slug reserved 30 days)",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "x-permission": {
          "action": "services.delete",
          "roles": [
            "owner",
            "admin"
          ],
          "apiKeyScope": null
        },
        "x-csrf": "Cookie-authenticated calls need Origin + X-CSRF-Token; bearer keys are exempt.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Team for this request (session principals). Defaults to the session’s active team."
          }
        ],
        "responses": {
          "204": {
            "description": "OK"
          },
          "401": {
            "description": "Not signed in / invalid key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (role, scope, CSRF, team)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/services/{id}/plan": {
      "post": {
        "summary": "Change the plan (billed)",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "x-permission": {
          "action": "services.plan",
          "roles": [
            "owner",
            "admin"
          ],
          "apiKeyScope": null
        },
        "x-csrf": "Cookie-authenticated calls need Origin + X-CSRF-Token; bearer keys are exempt.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Team for this request (session principals). Defaults to the session’s active team."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "plan": {
                    "type": "string",
                    "enum": [
                      "free",
                      "starter",
                      "pro",
                      "business"
                    ]
                  }
                },
                "required": [
                  "plan"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Service"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in / invalid key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "402": {
            "description": "Payment required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (role, scope, CSRF, team)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Unavailable / not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}/env": {
      "put": {
        "summary": "Replace environment variables",
        "tags": [
          "Services"
        ],
        "security": [
          {
            "session": []
          },
          {
            "apiKey": []
          }
        ],
        "x-permission": {
          "action": "services.env.write",
          "roles": [
            "owner",
            "admin"
          ],
          "apiKeyScope": "full"
        },
        "x-csrf": "Cookie-authenticated calls need Origin + X-CSRF-Token; bearer keys are exempt.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Team for this request (session principals). Defaults to the session’s active team."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "env": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/EnvVarInput"
                    }
                  }
                },
                "required": [
                  "env"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API keys get keys only (values blank).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EnvVar"
                      }
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in / invalid key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (role, scope, CSRF, team)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (Retry-After)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}